Corregido bug menor
He cambiado en la clase staff un tipado incorrecto. nuevo_sueldo era referenciado como un float, pero después se comprobaba como un int. Ahora se comprueba si es un float.

+++ staff.py
@@ -74,7 +74,7 @@ class Staff(Persona):
         nuevo_sueldo: float
             Valor nuevo del sueldo del empleado
         """
-        if type(nuevo_sueldo)==int:
+        if type(nuevo_sueldo)==float:
             self.__sueldo=nuevo_sueldo
         else:
             print('Sueldo no válido, prueba con un número entero')
